+++ cranelift/codegen/src/isa/mod.rs
@@ -285,7 +285,7 @@ pub trait TargetIsa: fmt::Display + Send + Sync {
     /// The `num_labeled_funcs` argument here is the number of functions which
     /// will be "labeled" or might have calls between them, typically the number
     /// of defined functions in the object file.
-    fn text_section_builder(&self, num_labeled_funcs: u32) -> Box<dyn TextSectionBuilder>;
+    fn text_section_builder(&self, num_labeled_funcs: usize) -> Box<dyn TextSectionBuilder>;
 
     /// The function alignment required by this ISA.
     fn function_alignment(&self) -> u32;
